+++ http-cache-reqwest/src/test.rs
@@ -182,7 +182,7 @@ async fn custom_cache_mode_fn() -> Result<()> {
     // Construct reqwest client with cache defaults and custom cache mode
     let client = ClientBuilder::new(Client::new())
         .with(Cache(HttpCache {
-            mode: CacheMode::Default,
+            mode: CacheMode::NoStore,
             manager: manager.clone(),
             options: HttpCacheOptions {
                 cache_key: None,
@@ -209,6 +209,46 @@ async fn custom_cache_mode_fn() -> Result<()> {
     let url = format!("{}/", &mock_server.uri());
     // To verify our endpoint receives the request rather than a cache hit
     client.get(url.clone()).send().await?;
+
+    // Check no cache object was created
+    let data = manager.get(&format!("{}:{}", GET, &Url::parse(&url)?)).await?;
+    assert!(data.is_none());
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn override_cache_mode() -> Result<()> {
+    let mock_server = MockServer::start().await;
+    let m = build_mock(CACHEABLE_PUBLIC, TEST_BODY, 200, 2);
+    let _mock_guard = mock_server.register_as_scoped(m).await;
+    let url = format!("{}/test.css", &mock_server.uri());
+    let manager = MokaManager::default();
+
+    // Construct reqwest client with cache defaults and custom cache mode
+    let client = ClientBuilder::new(Client::new())
+        .with(Cache(HttpCache {
+            mode: CacheMode::Default,
+            manager: manager.clone(),
+            options: HttpCacheOptions {
+                cache_key: None,
+                cache_options: None,
+                cache_mode_fn: None,
+                cache_bust: None,
+            },
+        }))
+        .build();
+
+    // Remote request and should cache
+    client.get(url.clone()).send().await?;
+
+    // Try to load cached object
+    let data = manager.get(&format!("{}:{}", GET, &Url::parse(&url)?)).await?;
+    assert!(data.is_some());
+
+    let url = format!("{}/", &mock_server.uri());
+    // To verify our endpoint receives the request rather than a cache hit
+    client.get(url.clone()).with_extension(CacheMode::NoStore).send().await?;
 
     // Check no cache object was created
     let data = manager.get(&format!("{}:{}", GET, &Url::parse(&url)?)).await?;